[klibc] libgcc functions: shift counts are type "int"
Per the gcc internals manual, shift counts arguments are "int"
(unclear if it is actually "int" or if it's "SImode" a.k.a. int32_t.
However, it doesn't matter for Linux-compatible architectures.)

+++ usr/klibc/libgcc/__ashldi3.c
@@ -5,9 +5,9 @@
 #include <stdint.h>
 #include <stddef.h>
 
-uint64_t __ashldi3(uint64_t v, uint64_t cnt)
+uint64_t __ashldi3(uint64_t v, int cnt)
 {
-	int c = (int)cnt & 31;
+	int c = cnt & 31;
 	uint32_t vl = (uint32_t) v;
 	uint32_t vh = (uint32_t) (v >> 32);
